API: Merge `EventPropEmitter` into `EventPropSetter`)
